added num_queues parameter to nic_list
implemented num_queues option for nics in nutanix_virtual_machine
objects. If num_queues is changed for an existing nic, the VM will be
restarted. If a new nic is added to a VM with a queue number specified,
the VM will not be restarted.

If any nic parameters besides num_queues are changed, the VM should not
restart.